rapi rapi controller
