+++ app/Http/Controllers/RiwayatController.php
@@ -23,28 +23,7 @@ class RiwayatController extends Controller
 //  ================= Get All And Get By Params =================
 //  =============================================================
 
-    public function getAll(){
-        // memvalidasi jika bukan params angkot_id or supir_id
-        if(request()->all()){
-            if(!request(['angkot_id', 'supir_id'])){
-                return response()->json([
-                    'status' => 'failed',
-                    'message' => 'params not available',
-                ], 400);
-            }
-        }
 
-        // fungsi method filter sebenar nya bukan ngequery tapi untuk memfilter sesuai params
-        // makanya aku taruh di models aja biar aku gak pusing
-        // untuk proses query tetap disini dan disimpan ke dalam $riwayat
-        $riwayat =  Riwayat::with('supir')->filter(request(['angkot_id', 'supir_id']))->get();
-
-        return response()->json([
-            'status' => 'success',
-            'message' => 'ok',
-            'data' => $riwayat
-        ], 200);
-    }
 
 
     // ==============================================================
@@ -60,36 +39,7 @@ class RiwayatController extends Controller
     // ==============================================================
     //  =================== Create New History ======================
     //  =============================================================
-    public function createHistory(Request $request){
-        try{
-            $user_id = $request->input('user_id');
-            $angkot_id = $request->input('angkot_id');
-            $jumlah_pendapatan = $request->input('jumlah_pendapatan');
-            $waktu_narik = $request->input('waktu_narik');
-            $selesai_narik = $request->input('selesai_narik');
-
-            $riwayat = new Riwayat();
-            $riwayat->user_id = $user_id;
-            $riwayat->angkot_id = $angkot_id;
-            $riwayat->jumlah_pendapatan = $jumlah_pendapatan;
-            $riwayat->waktu_narik = $waktu_narik;
-            $riwayat->selesai_narik = $selesai_narik;
-            $riwayat->save();
-
-            return response()->json([
-                'status' => 'success',
-                'message' => 'Data created',
-                'data' => $riwayat,
-            ], 201);
-        } catch (\Exception $e) {
-            //return error message
-            return response()->json([
-                'status' => 'failed',
-                'message' => $e,
-                'data' => [],
-                ], 409);
-            }
-    }
+    
 
     // ==============================================================
     //  ================ Update Data History By Id ==================
